new styles added
added new function to initialize 2 grids (contains value and styles)
Forms will refer to the styleGrid from now on to get style.
New styles used to identify which cells are user filled with #s 1-9, solution generated, or randomly generated. Randomly generated cells cannot be changed as it is the random puzzle to solve